+++ dokkusd/cli.py
@@ -48,6 +48,10 @@ def main() -> None:
         help="Environment Variables in JSON dictionary",
         default=os.getenv("DOKKUSD_ENVIRONMENT_VARIABLES_JSON"),
     )
+    deploy_parser.add_argument(
+        "--environmentvariablesprefixedby",
+        help="Any Environmental variables prefixed with this will be given to the Dokku app.",
+    )
 
     ### Destroy
     destroy_parser = subparsers.add_parser("destroy")
@@ -84,6 +88,12 @@ def main() -> None:
             print("Must specify app name!")
             sys.exit(-1)
 
+        env_vars = {}
+        if args.environmentvariablesprefixedby:
+            for name, value in os.environ.items():
+                if name.startswith(args.environmentvariablesprefixedby):
+                    env_vars[name[len(args.environmentvariablesprefixedby) :]] = value
+
         deploy = Deploy(
             directory=os.getcwd(),
             remote_user=args.remoteuser,
@@ -93,6 +103,7 @@ def main() -> None:
             http_auth_user=args.httpauthuser,
             http_auth_password=args.httpauthpassword,
             environment_variables_json_string=args.environmentvariablesjson,
+            environment_variables=env_vars,
         )
         deploy.go()